// src/types/ProfileType.ts
interface ProfileInterface {
  nik: string;
  nama: string;
  jk: string;
  alamat: string;
  hp: string;
  divisi: string;
  jabatan: string;
  sp: string;
  cuti_terakhir: string;
  cuti_sekarang: string;
  status_aktif: string;
  status_karyawan: string;
  tgl_masuk: string;
  tgl_keluar: string;
  durasi_kontrak: string;
  kode_absensi: string;
}

export type {ProfileInterface}